\name{maxlike-package}
\alias{maxlike-package}
\docType{package}
\title{
  Fit models of the probability of occurrence to presence-only data.
}
\description{
  This is the model that the MaxEnt software is trying to fit.
}
\details{
\tabular{ll}{
Package: \tab maxlike\cr
Type: \tab Package\cr
Version: \tab 0.0-1\cr
Date: \tab 2011-03-18\cr
License: \tab What license is it under?\cr
LazyLoad: \tab yes\cr
}

}
\keyword{ package }
