\name{maxlike-package}
\alias{maxlike-package}
\docType{package}
\title{
  Estimate the probability of occurrence using presence-only data.
}
\description{
  Estimate the probability of occurrence using presence-only data.
}
\details{
\tabular{ll}{
Package: \tab maxlike\cr
Type: \tab Package\cr
Version: \tab 0.0-1\cr
Date: \tab 2011-04-18\cr
License: \tab GPL (>=3) \cr
LazyLoad: \tab yes\cr
}

}
\keyword{ package }
